moved functions to Utils
