+++ src/events/presenceUpdate.ts
@@ -1,6 +1,5 @@
 import {
 	ActionRowBuilder,
-	Activity,
 	ButtonBuilder,
 	ButtonStyle,
 	Channel,
@@ -8,9 +7,9 @@ import {
 	Presence,
 	TextChannel,
 } from "discord.js";
-import { activityTrackerBlacklistDb, activityTrackerLogDb } from "../db";
 import { addDefaultEmbedFooter } from "../util/misc/embeds";
 import { config } from "../config";
+import { getStopedActivities, blacklistCheck, logTime, msToTimeString } from "../util/tracker/presence";
 
 export default async function presenceUpdate(oldPresence: Presence | null, newPresence: Presence) {
 	if (!config.logActivity) return;
@@ -45,47 +44,4 @@ export default async function presenceUpdate(oldPresence: Presence | null, newPr
 
 		await (channel as TextChannel)?.send({ embeds: [embed], components: [row] });
 	});
-}
-
-async function getStopedActivities(oldPresence: Presence | null, newPresence: Presence): Promise<Activity[]> {
-	// 0 = game activity
-	const oldActivities = oldPresence?.activities.filter((value) => value.type === 0);
-	const newActivities = newPresence.activities.filter((value) => value.type === 0);
-
-	const stopedActivities: Activity[] = [];
-	oldActivities?.forEach((element) => {
-		if (!newActivities.some((e) => e.name === element.name)) stopedActivities.push(element);
-	});
-
-	return stopedActivities;
-}
-
-async function blacklistCheck(userid: string, elementName: string): Promise<boolean> {
-	if (activityTrackerBlacklistDb.get("general-user")?.includes(userid)) return true;
-	if (activityTrackerBlacklistDb.get("general-game")?.includes(elementName)) return true;
-	if (activityTrackerBlacklistDb.has(userid)) {
-		if (activityTrackerBlacklistDb.get(userid)?.includes(elementName)) return true;
-	}
-	return false;
-}
-
-async function msToTimeString(ms: number): Promise<string> {
-	let totalSeconds = ms / 1000;
-	totalSeconds %= 86400;
-	const hours = Math.floor(totalSeconds / 3600);
-	totalSeconds %= 3600;
-	const minute = Math.floor(totalSeconds / 60);
-	const second = Math.floor(totalSeconds % 60);
-
-	return `${hours > 0 ? hours + "hour(s) " : ""}${hours > 0 && minute > 0 ? ", " : " "}${
-		minute > 0 ? minute + "minute(s) " : ""
-	}${(hours > 0 || minute > 0) && second > 0 ? "and " : ""}${
-		second > 0 ? second + " second(s) " : ""
-	}`.trim();
-}
-
-async function logTime(userid: string, elementName: string, timePlayed: number): Promise<void> {
-	if (!activityTrackerLogDb.has(`${userid}-${elementName}`))
-		activityTrackerLogDb.set(`${userid}-${elementName}`, []);
-	activityTrackerLogDb.push(`${userid}-${elementName}`, { t: timePlayed, w: Date.now() });
 }
